requirements.txt
click==6.7
azure-storage==0.34.3
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10.1
MarkupSafe==1.0
Werkzeug==0.14.1
json2html==1.3.0
beautifulsoup4==4.7.1
lxml==4.3.3
applicationinsights==0.11.3
pandas==0.24.2
